feat: use ariaNotify when available
