+++ packages/react-components/react-aria/src/AriaLiveAnnouncer/useAriaNotifyAnnounce.ts
@@ -1,0 +1,46 @@
+import { useFluent_unstable as useFluent } from '@fluentui/react-shared-contexts';
+import * as React from 'react';
+
+import type { AriaLiveAnnounceFn } from './AriaLiveAnnouncer.types';
+
+type AriaNotifyOptions = {
+  notificationID?: string;
+  priority?: 'none' | 'important';
+  interrupt?: 'all' | 'pending' | 'none';
+};
+
+type DocumentWithAriaNotify = Document & {
+  ariaNotify: (message: string, options: AriaNotifyOptions) => void;
+};
+
+/* INTERNAL: implementation of the announcer using the ariaNotify API */
+export const useAriaNotifyAnnounce_unstable = (): AriaLiveAnnounceFn => {
+  const { targetDocument } = useFluent();
+
+  const announce: AriaLiveAnnounceFn = React.useMemo(
+    () =>
+      (message, options = {}) => {
+        if (!targetDocument) return;
+
+        let { alert = false, polite, priority, batchId } = options;
+
+        // default priority to 0 if polite, 2 if alert, and 1 by default
+        // used to set both ariaNotify's priority and interrupt
+        if (typeof priority !== 'number') {
+          priority = polite ? 0 : alert ? 2 : 1;
+        }
+
+        // map fluent announce options to ariaNotify options
+        const ariaNotifyOptions: AriaNotifyOptions = {
+          notificationID: batchId,
+          priority: priority > 1 ? 'important' : 'none',
+          interrupt: batchId ? (priority > 0 ? 'all' : 'pending') : 'none',
+        };
+
+        (targetDocument as DocumentWithAriaNotify).ariaNotify(message, ariaNotifyOptions);
+      },
+    [],
+  );
+
+  return announce;
+};
